Pull in updated realtime types

+++ lib/mux_ruby/models/real_time_histogram_timeseries_datapoint.rb
@@ -39,8 +39,8 @@ module MuxRuby
       {
         :'timestamp' => :'String',
         :'sum' => :'Integer',
-        :'p95' => :'Integer',
-        :'median' => :'Integer',
+        :'p95' => :'Float',
+        :'median' => :'Float',
         :'max_percentage' => :'Float',
         :'bucket_values' => :'Array<RealTimeHistogramTimeseriesBucketValues>',
         :'average' => :'Float'
